Introduce release script
The script is fairly simple for now. Later I want to have it read
CHANGELOG.md and produce a nice git tag/release based on the changelog
information for the specific version. This way we can also enforce
that you can't release without having a changelog entry for the
specific release :)